requirements.txt
asgiref==3.8.1
django==5.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.0
typing-extensions==4.11.0
